Simple timeline
This commit implements a simple instagram-like timeline with photo uploads and
filters.

Instagram core experience is checking the timeline and uploading new pictures,
so it makes sense to explore the app with this feature. That way I can
understand the problem better and check whether there are any unexpected
obstacles.

Notice that I haven’t implemented users and authentication. For the time being
there is only one shared timeline with all the posts in the app. This is to
start the project with from the core feature, but I’ll fix this in another PR
that adds users to the application.

The timeline works with a simple `Post` model that has a picture and a caption.
I’m also storing the original picture upload (before any filters are applied)
just in case we need to come back to the select filter screen.

The upload picture screen works in two screens, pretty much like the Instagram
app.

In the first screen you can select a filter. There’s no cropping at the moment,
but I’ll add that later.

In the second screen you add the caption. In this screen eventually you’ll be
able to add other metadata like location.

As for the filters, they are implemented client side with
the[instagram-filters](https://gitlab.com/lebrun.noe/instagram-filters) npm
package. Since I can not install the package with yarn because I’m not using
webpacker, I’m importing the package directly from the web using
[Skypack](https://cdn.skypack.dev).

There are three small stimulus controllers:

* `form_controller.js` submits the upload form when a photo is selected.
* `preset_controller.js` applies a filter preset to a thumbnail preview.

* Finally, `picture_controller.js` updates the uploaded picture when a filter is
  selected, serialises the image in base64 and stores the result in a hidden
input field. That input field with the base64 image representation is sent to
the server when the user moves to the next screen.

On the server side, there are three rails controllers, one for each of the
screens implemented (timeline + the 2 upload picture  screens). Notice that all
these controllers operate on the same `Post` model, but I find it useful to
split the different operations into separate controllers so the remain small and
easy to understand.